Project.1.DownloadFromYoutube/requirements.txt
beautifulsoup4==4.9.0
html5lib==1.0.1
mechanize==0.4.5
pytube3==9.6.4
six==1.14.0
soupsieve==2.0
typing-extensions==3.7.4.2
webencodings==0.5.1
